% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate_lavaan.R
\name{cross_validate_lavaan}
\alias{cross_validate_lavaan}
\title{cross_validate_lavaan}
\usage{
cross_validate_lavaan(lavaan_model, test_set)
}
\arguments{
\item{lavaan_model}{model fitted with lavaan}

\item{test_set}{raw data set used for cross-validation}
}
\value{
-2 log-Likelihood of the test-set
}
\description{
Cross-validate a model fitted with lavaan using a separate test set. Currently
only supports single-group models fitted with (full information) maximum likelihood.
}
\examples{
# example code

}
