% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_specification_search.R
\name{plot.Boot_Spec_Search}
\alias{plot.Boot_Spec_Search}
\title{plot.Boot_Spec_Search}
\usage{
\method{plot}{Boot_Spec_Search}(x, y = NULL, ...)
}
\arguments{
\item{x}{object of class Boot_Spec_Search}

\item{y}{not used}

\item{...}{not used}
}
\value{
ggplot2 object
}
\description{
Plot the results of a bootstrap specification search
}
