% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{semmi}
\alias{semmi}
\alias{semmi-package}
\title{semmmi}
\description{
Modification indices are commonly used in structural equation models (SEM) to post-hoc improve the fit. This practice has been criticized for capitalization on chance (i.e., overfitting). semmi (SEM modification indices) provides bootstrapped or cross-validated modification indices to evaluate the stability of added modifications.
}
\author{
Jannik Orzek \href{mailto:orzek@mpib-berlin.mpg.de}{orzek@mpib-berlin.mpg.de}
}
\keyword{internal}
