% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_specification_search.R
\name{show.Boot_Spec_Search}
\alias{show.Boot_Spec_Search}
\title{show.Boot_Spec_Search}
\usage{
\method{show}{Boot_Spec_Search}(object)
}
\arguments{
\item{object}{object of class Boot_Spec_Search}
}
\value{
tibble with summarized results
}
\description{
show results of specification_search
}
