% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_specification_search.R
\name{show.resample_Spec_Search}
\alias{show.resample_Spec_Search}
\title{show.resample_Spec_Search}
\usage{
\method{show}{resample_Spec_Search}(object)
}
\arguments{
\item{object}{object of class resample_Spec_Search}
}
\value{
tibble with summarized results
}
\description{
show results of specification_search
}
