% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification_search.R
\name{show.Spec_Search}
\alias{show.Spec_Search}
\title{show.Spec_Search}
\usage{
\method{show}{Spec_Search}(object)
}
\arguments{
\item{object}{object of class Spec_Search}
}
\value{
nothing
}
\description{
show results of specification_search
}
