% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_PolDem}
\alias{simulate_PolDem}
\title{simulate_PolDem}
\usage{
simulate_PolDem(sample_size = 100)
}
\arguments{
\item{sample_size}{sample size}
}
\value{
data and model used to simulate the data
}
\description{
Simulate example data based on the PoliticalDemocracy data set.
}
\examples{
set.seed(2343)
data <- simulate_PolDem()
}
