% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_specification_search.R
\name{summary.Boot_Spec_Search}
\alias{summary.Boot_Spec_Search}
\title{summary.Boot_Spec_Search}
\usage{
\method{summary}{Boot_Spec_Search}(object, ...)
}
\arguments{
\item{object}{object of class Boot_Spec_Search}

\item{...}{not used}
}
\value{
tibble with summarized results
}
\description{
show results of specification_search
}
