% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_specification_search.R
\name{summary.resample_Spec_Search}
\alias{summary.resample_Spec_Search}
\title{summary.resample_Spec_Search}
\usage{
\method{summary}{resample_Spec_Search}(object, ...)
}
\arguments{
\item{object}{object of class resample_Spec_Search}

\item{...}{not used}
}
\value{
tibble with summarized results
}
\description{
show results of specification_search
}
