% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification_search.R
\name{summary.Spec_Search}
\alias{summary.Spec_Search}
\title{summary.Spec_Search}
\usage{
\method{summary}{Spec_Search}(object, ...)
}
\arguments{
\item{object}{object of class Spec_Search}

\item{...}{not used}
}
\value{
nothing
}
\description{
show results of specification_search
}
